Arnold (Arnold v0.6.2)

Arnold main API functions, which can be used to communicate with other nodes or applications. The REST Api utilizes these functions mainly.

Functions

analyse(node_id, sensor_id, tag, prediction)
(since 0.6.2)

@spec analyse(
  node_id :: binary(),
  sensor_id :: binary(),
  tag :: Arnold.Sensor.tag(),
  prediction :: Arnold.NeuralNetwork.prediction()
) :: {nil, nil} | {msg :: binary(), map()}

Returns the correlation with each metric that was pushed for a node. If found positive or negative correlation for a threshold larger (or smaller) than 0.7 marks it as the metrics are correlating with each other.

Example

iex(1)> Arnold.analyze("node", "sensor_id", :hourly, [[1642433780, 1642433840, 1642433900, 1642433960,1642434020],
                                                      [119027224.0,119214032.0, 119363472.0, 119483024.0, 119578672.0],
                                                      [107124496.0, 107292624.0, 107427120.0, 107534720.0, 1.076208e8]
                                                      [130929952.0, 131135440.0, 131299824.0, 131431328.0, 131536544.0]])
{:alarm, "Current value is higher than the expected prediction range by X"}
iex(2)>

feed(node_id, sensor_id, time, value)
(since 0.5.4)

@spec feed(
  node_id :: binary(),
  sensor_id :: binary(),
  time :: pos_integer(),
  value :: number()
) :: :ok

Feed data to the sensors. It is going to be marked for train if the threshold is reached for the current input category.

Example

iex(1)> Arnold.feed("node", "sensor_id", 1642433780, 5)
:ok
iex(2)>

predict(id, tag, horizon)
(since 0.5.4)

@spec predict(id :: binary(), tag :: Arnold.Sensor.tag(), horizon :: pos_integer()) ::
  {:ok, Arnold.NeuralNetwork.prediction()}

Predicts a value for a given id and tag. Check out the :windows to see the available tag values. There are two types of predictions: simple and complex. Complex uses the trained neural network, while the simple is only used when there aren't any networks for the current id. It uses SES algorithm for predictions.

Example

iex(1)> Arnold.predict("node_sensor_uuid", :hourly, 5)
{ok,
    [[1642433780, 1642433840, 1642433900, 1642433960,1642434020],
     [119027224.0,119214032.0, 119363472.0, 119483024.0, 119578672.0],
     [107124496.0, 107292624.0, 107427120.0, 107534720.0, 1.076208e8]
     [130929952.0, 131135440.0, 131299824.0, 131431328.0, 131536544.0]]}
iex(2)>

predict(node_id, sensor_id, tag, horizon)
(since 0.6.2)

@spec predict(
  node_id :: binary(),
  sensor_id :: binary(),
  tag :: Arnold.Sensor.tag(),
  horizon :: pos_integer()
) :: {:ok, Arnold.NeuralNetwork.prediction()}

Predicts a value for a given node, sensor_id and tag. Check out the :windows to see the available tag values. Calls for predict/2.

Example

iex(1)> Arnold.predict("node", "sensor_id", :hourly, 5)
{ok,
    [[1642433780, 1642433840, 1642433900, 1642433960,1642434020],
     [119027224.0,119214032.0, 119363472.0, 119483024.0, 119578672.0],
     [107124496.0, 107292624.0, 107427120.0, 107534720.0, 1.076208e8]
     [130929952.0, 131135440.0, 131299824.0, 131431328.0, 131536544.0]]}
iex(2)>