Arnold.Config (Arnold v0.6.2)

Configuration handler module for Arnold.

Functions

get(key)
(since 0.6.2)

@spec get(key :: atom()) :: {:ok, value :: [{atom(), pos_integer()}] | number()}

Fetches a config value with a given key. The following default configs are configured for Arnold:

Example

iex(1)> Arnold.Config.get(:port)
{:ok, 8081}
iex(2)>

The windows and ports are hard coded with the following default values:

:hourly = 60
:daily = 96
:weekly = 168
:port = 8081

Port can be configured via a config file or setting the ARNOLD_PORT environment variable.

Example

iex(1)> Arnold.Config.get(:window, :hourly)
{:ok, 60, 3600}
iex(2)>

get(atom1, atom2)
(since 0.5.4)

@spec get(env :: atom(), window_type :: atom()) :: {:ok, pos_integer(), pos_integer()}

Returns a value for a given window_type. The following types are accepted:

:hourly

:daily

:weekly

Example

iex(1)> Arnold.Config.get(:window, :hourly)
{:ok, 60, 3600}
iex(2)>