Arnold.Database (Arnold v0.6.2)

API wrapper for Memento to store or fetch data from the database.

Functions

get(table, uuid)

@spec get(schema :: module(), uuid :: binary()) ::
  {:ok, term()} | {:error, :not_found}

Fetches data from database. Needs a table schema and an uuid which is made from the combination of node_name and sensor_id with Arnold.Utilities.id/2.

Example

iex(1)> Arnold.Database(Arnold.Database.Table.Sensor, "85a68cb6-04c7-5a78-8f4c-bcc184186b6b")
{:ok, %Arnold.Database.Table.Sensor{
  id: "85a68cb6-04c7-5a78-8f4c-bcc184186b6b",
  hourly: [{1610986020, 5}],
  daily: [],
  weekly: [],
  predictions: %{hourly: [[],[],[],[]], daily: [[],[],[],[]], weekly: [[],[],[],[]]}
}}
iex(2)>

get(schema, node_id, sensor_id)

@spec get(schema :: module(), node_id :: binary(), sensor_id :: binary()) ::
  {:ok, term()} | {:error, :not_found}

Tries to fetch a specific node_id and sensor_id from a table. If not found returns all the currently saved data from the backend

insert(data)

@spec insert(data :: list() | map()) :: :ok

Inserts the data to the specific table of the Postgres database. If it's there updates the value