Arnold.Manager (Arnold v0.6.2)

The manager responsible for saving sensors state, finished sensor tags and the number of sensor agents. If an agent terminates it the duty of the module to start a new one.

Functions

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

get_metrics(node)
(since 0.6.2)

@spec get_metrics(node :: binary()) :: MapSet.t()

Returns all metrics for a given node. Used for getting all the metrics for a given node to check the correlations. Returns an empty MapSet if node not found.

Example

iex(1)> Arnold.Manager.get("node")
#MapSet<["sensor_id"]>
iex(2)>

get_state()
(since 0.5.4)

@spec get_state() :: map()

Returns the state of the manager server.

Example

iex(1)> Arnold.Manager.get_state
%{:timers =>
  %{sensor: {:interval, #Reference<0.3386470258.2604662794.215684>},
    marks: {:interval, #Reference<0.3386470258.2604662794.315674>}},
  :marks => #MapSet<[]>,
  :finished => #MapSet<[]>,
  :metriccs => Map}
iex(2)>

is_finished?(id, tag)
(since 0.5.4)

@spec is_finished?(id :: binary(), tag :: Arnold.Sensor.tag()) :: boolean()

Checks if the id with the given tag has already finished training.

Example

iex(1)> Arnold.Manager.is_finished?("node_sensor_id", :hourly)
true
iex(2)>

mark(id, tag)
(since 0.5.4)

@spec mark(id :: binary(), tag :: Arnold.Sensor.tag()) :: :ok

Marks an id with a given tag for training.

Example

iex(1)> Arnold.Manager.mark("node_sensor_id", :hourly)
:ok
iex(2)>

register_metric(node, sensor_id)
(since 0.6.2)

@spec register_metric(node :: binary(), sensor_id :: binary()) :: :ok

Registers a metric to the state of the manager. Later is used for getting all the metrics for a given node to check the correlations.

Example

iex(1)> Arnold.Manager.register("node", "sensor_id")
:ok
iex(2)>

start_link(state \\ %{}, opts \\ [])

Starts a GenServer process linked to the current process. Check GenServer.start_link/3 for more.