Arnold.NeuralNetwork (Arnold v0.6.2)

The neural network API that contains function for training and prediction.

Types

prediction()
(since 0.6.2)

@type prediction() :: [[integer()]]

Prediction result. A 4 element list of list of integers.

List of timestamps

List of predictions

List of upper range predictions

List of lower range predictions

[[],[],[],[],[]]

prediction_method()
(since 0.5.4)

@type prediction_method() :: :simple | :complex

Prediction method type that can be set as the third argument of Arnold.NeuralNetwork.predict/4

Functions

predict(sensor, tag, horizon, method)
(since 0.6.2)

@spec predict(
  sensor :: Arnold.Database.Table.Sensor.t() | nil,
  tag :: Arnold.Sensor.tag(),
  horizon :: pos_integer(),
  type :: prediction_method()
) :: {:ok, prediction :: prediction()}

Predict a value or a list of values or a single value based on the method which can be :complex or :simple. Simple type utilizes the Single Exponential Smoothing algorithm from Arnold.Analyser.ExponentialSmoothing.simple/1, which is only good for predicting one value ahead. Complex uses neural network while also affects the results by trend and seasonality.

If the sensor is nil it returns nil.

Example

iex(1)> Arnold.NeuralNetwork.predict(%Arnold.Database.Table.Sensor{
...(1)> id: "85a68cb6-04c7-5a78-8f4c-bcc184186b6b",
...(1)> hourly: [{1642594108, 4}], daily: [], weekly: []}, :hourly, 1, :simple)
{:ok, [[1642594108], [4], [4.8], [3.2]]}
iex(2)>

save(neural_network_data)

@spec save(Arnold.Database.Table.NetworkModel.t()) :: :ok

Saves the given result from the Arnold.NeuralNetwork.train/1 function.

train(sensor)
(since 0.6.0)

@spec train(sensor :: Arnold.Database.Table.Sensor.t()) ::
  Arnold.Database.Table.NetworkModel.t()

Trains the neural network with the given sensor and tag, until a configure eps value is not reached. If it fails to reach the given epsilon value, the training stops when exceeding the :max_iterates config value.

Example

iex(1)> Arnold.NeuralNetwork.train(%Arnold.Database.Table.Sensor{
...(1)> id: "85a68cb6-04c7-5a78-8f4c-bcc184186b6b",
...(1)> hourly: [{1642593268,5},{1642593208, 6},{1642593148, 4},{1642593088, 8},{1642593028, 0}...],
...(1)> daily: [{1642594108, 4},{1642593808, 7},{1642593508, 2},{1642593208, 5}], weekly: []}, :hourly)
%Arnold.Database.Table.NetworkModel{}
iex(2)>