Arnold.Plug.Router (Arnold v0.6.2)

Router for the REST API. There are 2 main routes, get and post paths. The body must be a json in all cases.

GET paths:

/api/tendency (not used)

/api/seasonality (not used)

/api/prediction

All of them needs 3 common parameters:

node

metric

tag

Prediction route needs a 4th one:

horizon

As a result a sample url path should look like this (port can be customized): http://localhost:8081/api/tendency?node=node_id&metric=metric_id&tag=hourly

POST path:

/api/write

Parameters:

node

metric

Body:

type

value

timestamp Currently only gauge, counter, meter, spiral, histogram and durations are accepted via the router. A sample url: http://localhost:8081/api/write?node=node_id&metric=metric_id with body as json

{
  "type": "gauge",
  "timestamp": 1642433780,
  "value": 75
}

Both GET and POST methods has built-in request verification. Arnold.Plug.InvalidGetParams or Arnold.Plug.InvalidPostParams errors are raised if the sent request are not valid.

Functions

call(conn, opts)

Callback implementation for Plug.call/2.

init(opts)

Callback implementation for Plug.init/1.