Arnold.Sensor (Arnold v0.6.2)

Contains the API functions for the sensors

Types

tag()
(since 0.6.2)

@type tag() :: :hourly | :daily | :weekly

Time tag for sensor

Functions

all()
(since 0.6.0)

@spec all() :: [Arnold.Database.Table.Sensor.t()]

Returns all sensors from every sensor agent.

Example

iex(1)> Arnold.Sensor.all
[%Arnold.Database.Table.Sensor{
    id: "85a68cb6-04c7-5a78-8f4c-bcc184186b6b",
    hourly: [{1610986020, 5}],
    daily: [],
    weekly: [],
    predictions: %{hourly: [[],[],[],[]], daily: [[],[],[],[]], weekly: [[],[],[],[]]}
},
%Arnold.Database.Table.Sensor{
    id: "adff7462-a100-55be-8ee1-8f8c34d3a9e3",
    hourly: [{1610986020, 6}],
    daily: [],
    weekly: [],
    predictions: %{hourly: [[],[],[],[]], daily: [[],[],[],[]], weekly: [[],[],[],[]]}
}]
iex(2)>

delete(uuid)
(since 0.5.4)

@spec delete(uuid :: binary()) :: :ok

Deletes a sensor from a given uuid. UUID is the combination of node_id and sensor_id made with Arnold.Utilities.id/2.

Example

iex(1)> Arnold.Sensor.delete("85a68cb6-04c7-5a78-8f4c-bcc184186b6b")
:ok
iex(2)>

get(uuid)
(since 0.6.0)

@spec get(id :: binary()) :: Arnold.Database.Table.Sensor.t() | nil

Returns a sensor from with the given uuid. Only returns the sensor if it is in the state of the agent. UUID is the combination of node_id and sensor_id made with Arnold.Utilities.id/2.

Example

iex(1)> Arnold.Sensor.get("85a68cb6-04c7-5a78-8f4c-bcc184186b6b")
%Arnold.Database.Table.Sensor{
    id: "85a68cb6-04c7-5a78-8f4c-bcc184186b6b",
    hourly: [{1610986020, 5}],
    daily: [],
    weekly: [],
    predictions: %{hourly: [[],[],[],[]], daily: [[],[],[],[]], weekly: [[],[],[],[]]}
}
iex(2)>

merge(sensor, tag, values)

new(uuid, timestamp, value)
(since 0.6.2)

@spec new(uuid :: binary(), timestamp :: pos_integer(), value :: number()) ::
  Arnold.Database.Table.Sensor.t()

Creates a new Sensor with the given uuid and input value. UUID is the combination of node_id and sensor_id made with Arnold.Utilities.id/2.

Example

iex(1)> Arnold.Sensor.new("85a68cb6-04c7-5a78-8f4c-bcc184186b6b", 1610986020, 5)
%Arnold.Database.Table.Sensor{
    id: "85a68cb6-04c7-5a78-8f4c-bcc184186b6b",
    hourly: [{1610986020, 5}],
    daily: [],
    weekly: [],
    predictions: %{hourly: [[],[],[],[]], daily: [[],[],[],[]], weekly: [[],[],[],[]]}
  }
iex(2)>

put(node_id, sensor_id, timestamp, value)
(since 0.6.0)

@spec put(
  node_id :: binary(),
  sensor_id :: binary(),
  time :: pos_integer(),
  value :: number()
) ::
  Arnold.Database.Table.Sensor.t()

Puts a given sensor with a value into one of the agent's state. The function creates a new sensor if the agent failed to find it in its state or in the database, or if found one updates it.

The agent calls the Arnold.Sensor.new/3 and Arnold.Sensor.update/3 functions.

Example

iex(1)> Arnold.Sensor.put("node", "sensor", 5)
%Arnold.Database.Table.Sensor{
    id: "85a68cb6-04c7-5a78-8f4c-bcc184186b6b",
    hourly: [{1610986020, 5}],
    daily: [],
    weekly: [],
    predictions: %{hourly: [[],[],[],[]], daily: [[],[],[],[]], weekly: [[],[],[],[]]}
  }
iex(2)>

reset(uuid, tag)
(since 0.5.4)

@spec reset(uuid :: binary(), tag :: tag()) :: :ok

Resets a sensor with a given uuid and tag. It means that truncates the input values for the given tag. UUID is the combination of node_id and sensor_id made with Arnold.Utilities.id/2.

Example

iex(1)> Arnold.Sensor.reset("85a68cb6-04c7-5a78-8f4c-bcc184186b6b", :hourly)
:ok
iex(2)>

update(sensor, timestamp, value)
(since 0.6.0)

@spec update(sensor :: map(), timestamp :: pos_integer(), value :: number()) :: map()

Updates a given sensor with a value

Example

iex(1)> Arnold.Sensor.update(%Arnold.Database.Table.Sensor{
...(1)> id: "85a68cb6-04c7-5a78-8f4c-bcc184186b6b",
...(1)> hourly: [{1610986020, 5}],
...(1)> daily: [],
...(1)> weekly: [],
...(1)> predictions: %{hourly: [[],[],[],[]], daily: [[],[],[],[]], weekly: [[],[],[],[]]}}, 1642593088, 7)
%Arnold.Database.Table.Sensor{
    id: "85a68cb6-04c7-5a78-8f4c-bcc184186b6b",
    hourly: [{1642593088, 7},{1610986020, 5}],
    daily: [],
    weekly: [],
    predictions: %{hourly: [[],[],[],[]], daily: [[],[],[],[]], weekly: [[],[],[],[]]}
  }
iex(2)>

values(input)

@spec values(input :: [tuple()]) :: [number()]

Gets the values from the input proplist.

Example

iex(1)> Arnold.Sensor.values([{1610986020, 46386}, {1610989620, 3456}])
[46386, 3456]
iex(2)>