Arnold.Statistics.Decomposition (Arnold v0.6.2)

Decompose the input values into trend and seasonality.

Functions

execute(input)
(since 0.5.4)

@spec execute(input :: Nx.Tensor.t() | list()) ::
  {Nx.Tensor.t(), [Nx.Tensor.t()], Nx.Tensor.t()}

Decompose uses a multiplicative decomposition method.

InputData = Trend * Seasonality * Noise The function returns a 3-element tuple, which contains where the the first element is the trend, followed by seasonality, then noise for the last place.

For more information checkout this page: https://otexts.com/fpp2/components.html

trend(input)