Arnold.Statistics.Math (Arnold v0.6.2)

Collection of commonly used mathematical/statistic functions.

Functions

devsq(x)
(since 0.5.4)

@spec devsq(x :: Nx.Tensor.t()) :: Nx.Tensor.t()

Calculates the sum of the squared deviations from the mean, without dividing by N or by N-1.

Example

iex(1)> Nx.tensor([1,2,3,4,5,6]) |> Arnold.Analyser.Math.devsq
#Nx.Tensor<
  f32
  17.5
>
iex(2)>

differencing(x, d, window \\ 2)

percentage(x, y)
(since 0.5.4)

@spec percentage(x :: number(), y :: number()) :: number()

Calculates the percentage between 0 and 1 of x and y. Always the smaller number is divided by the larger.

Example

iex(1)> Arnold.Analyser.Math.percentage(5,4)
0.8
iex(2)> Arnold.Analyser.Math.percentage(4,5)
0.8
iex(3)>

sma(input, window)
(since 0.5.4)

@spec sma(x :: Nx.Tensor.t(), window :: pos_integer()) :: Nx.Tensor.t()

Simple Moving Average algorithm

Calculation that is used to analyze data points by creating a series of averages of different subsets of the full data set.

Example

iex(1)> Nx.tensor([1,2,3,4,5,6]) |> Arnold.Analyser.Math.sma(2)
#Nx.Tensor<
  f32[5]
  [1.5, 2.5, 3.5, 4.5, 5.5]
>
iex(2)>

std(x, opts \\ [])
(since 0.6.0)

@spec std(x :: Nx.Tensor.t(), opts :: list()) :: Nx.Tensor.t()

Function for calculating the standard deviation of the given set of numbers.

Example

iex(1)> Arnold.Analyser.Math.std(Nx.tensor([4,7,2,1,6,3]))
#Nx.Tensor<
  f32
  2.114763021469116
>
iex(2)>

variance(x)
(since 0.5.4)

@spec variance(x :: Nx.Tensor.t()) :: Nx.Tensor.t()

Estimates the variance of a sample of data.

The variance function is a measure of heteroscedasticity and plays a large role in many settings of statistical modelling

Example

iex(1)> Nx.tensor([1,2,3,4,5,6]) |> Arnold.Analyser.Math.variance
#Nx.Tensor<
  f32
  3.5
>
iex(2)>