Arnold.Utilities (Arnold v0.6.2)

Collection of utility functions including id generation, truncate and normalization functions.

Functions

denormalize(x, min, max)
(since 0.5.4)

@spec denormalize(x :: number(), min :: number(), max :: number()) :: float()

Denormalizes a value with the given min and max value.

Example: iex(1)> Arnold.Utilities.denormalize(0.75,1,5) 4.0 iex(2)>

id(id1, id2)
(since 0.6.2)

@spec id(id1 :: binary(), id2 :: binary()) :: binary()

Return a generated UUID v5 based on the given ids.

Example

iex(1)> Arnold.Utilities.id("node", "sensor")
"85a68cb6-04c7-5a78-8f4c-bcc184186b6b"
iex(2)>

integer_to_boolean(number)

is_constant?(list)
(since 0.5.4)

@spec is_constant?(x :: list()) :: boolean()

Checks if the given list is only has constant values.

Example

iex(1)> Arnold.Utilities.is_constant?([1,2,3,4,5,6])
false
iex(2)>

normal_result(result_of_normalize_function)
(since 0.5.4)

@spec normal_result({x :: number() | list(), min :: number(), max :: number()}) ::
  number() | [number()]

Returns the result of the normalize function return value. Mainly used in pipe operations

Example

iex(1)> Arnold.Utilities.normalize(4,1,5)
...(1)> |> Arnold.Utilities.normal_result
0.75
iex(2)>

normalize(inputs)
(since 0.5.4)

@spec normalize(inputs :: [number()]) :: {[number()], integer(), integer()}

Normalizes a list of floats or integers with min-max normalization. Automatically gets the minimum and maximum values from the given list. Calls to Arnold.Utilities.normalize/3

Example

iex(1)> Arnold.Utilities.normalize([1,2,3,4,5])
{[0.0, 0.25, 0.5, 0.75, 1.0], 1, 5}
iex(2)>

normalize(inputs, min, max)
(since 0.5.4)

@spec normalize(x :: number() | list(), min :: number(), max :: number()) ::
  {number() | [number()], integer(), integer()}

Normalizes a list of floats or integers, or single values with min-max normalization. For every element Arnold.Utilities.normalize/3 is called.

Example

iex(1)> Arnold.Utilities.normalize(4,1,5)
{0.75, 1, 5}
iex(2)>

truncate(inputs, threshold)
(since 0.5.4)

@spec truncate(list(), pos_integer()) :: list()

Truncates a given list to the given length

Example

iex(1)> Arnold.Utilities.truncate([1,2,3,4,5], 4)
[1, 2, 3, 4]
iex(2)>